🏗️ Overhauls Bash/Zsh Profiles
This overhauls the Bash/Zsh profiles so that they have expanded capability while having as much feature parity between Bash and Zsh as reasonably possible. This new design is particularly modular, to make it easy to quickly add new bits of functionality in the future.

The Bash/Zsh profiles are now also somewhat covered with tests, which was a necessary (but fortunate) side-effect of having to iterate on them so heavily as part of this refactor.